Move LibraryDefinitionsBuilder into separate file
